containers: do not run yum in Ubuntu

+++ contrail-containers/common/aws/create-vm.sh
@@ -131,13 +131,16 @@ function run_instance() {
     echo "WARNING: Machine isn't accessible yet"
     sleep 2
   done
-  $ssh "sudo yum -y update"
-  $ssh "sudo reboot" || /bin/true
-  echo "INFO: reboot & waiting for instance SSH"
-  while ! $ssh uname -a 2>/dev/null ; do
-    echo "WARNING: Machine isn't accessible yet"
-    sleep 2
-  done
+  if [[ "$ENVIRONMENT_OS" == 'centos' ]]; then
+    # there are some cases when AWS image has strange kernel version and vrouter can't be loaded
+    $ssh "sudo yum -y update"
+    $ssh "sudo reboot" || /bin/true
+    echo "INFO: reboot & waiting for instance SSH"
+    while ! $ssh uname -a 2>/dev/null ; do
+      echo "WARNING: Machine isn't accessible yet"
+      sleep 2
+    done
+  fi
   if [[ $cloud_vm == "true" ]]; then
     echo "INFO: Configure additional disk for cloud VM"
     $ssh "(echo o; echo n; echo p; echo 1; echo ; echo ; echo w) | sudo fdisk /dev/xvdf"
